view individual messages + bump version
