+++ app/routes/message.js
@@ -1,18 +1,11 @@
 const { GET } = require('../constants/http-verbs')
 const Wreck = require('@hapi/wreck')
 const { serverConfig } = require('../config')
-
-const formatDate = (dateString) => {
-  const date = new Date(dateString)
-  const day = date.getDate()
-  const month = date.toLocaleDateString('default', { month: 'long' })
-  const year = date.getFullYear()
-  return `${day} ${month} ${year}`
-}
+const { formatDate } = require('../utils/format-date')
 
 module.exports = {
   method: GET,
-  path: '/{id}',
+  path: '/message/{id}',
   handler: async (request, h) => {
     try {
       const messageId = request.params.id
@@ -20,11 +13,13 @@ module.exports = {
         `${serverConfig.messagesHost}/messages/${messageId}`,
         { json: true }
       )
-      const message = {
+      const messageData = {
         ...response.payload,
         requestedDate: formatDate(response.payload.requestedDate)
       }
-      return h.view('message-view', { message })
+
+      console.log(messageData)
+      return h.view('message', { messageData })
     } catch (error) {
       console.error(error)
     }
