+++ app/views/message.njk
@@ -40,34 +40,25 @@
     <div class="govuk-grid-column-full">
 
       <span class="govuk-caption-l">Email from:</span>
-
-      {% for message in messages %}
-        {% if message.messageId == messageId %}
-
-        <h1 class="govuk-heading-l">{{ message.scheme }}</h1>
+        <h1 class="govuk-heading-l">{{ messageData.scheme }}</h1>
         
-
-        <p class="govuk-body-s govuk-!-margin-bottom-1">Sent {{ message.requestedDate }}</p>
+        <p class="govuk-body-s govuk-!-margin-bottom-1">Sent {{ messageData.requestedDate }}</p>
         <div class="govuk-summary-card">
           <div class="govuk-summary-card__title-wrapper">
             <h2 class="govuk-summary-card__title">
-              {{ message.content.heading }}
+              {{ messageData.content.heading }}
             </h2>
           </div>
           <div class="govuk-summary-card__content">
             <p class="govuk-body">Claim number: AHWR 1234 5678</p>
-            <p class="govuk-body">{{ message.content.body }}</p>
+            <p class="govuk-body">{{ messageData.content.body }}</p>
 
-        {% endif %}
-      {% endfor %}
-
-          <h2 class="govuk-heading-s">What happens next</h2>
-          <p class="govuk-body">The Rural Payments Agency (RPA) may contact you to request the review summary the vet has given you.</p>
-    
-          <h2 class="govuk-heading-s">Get help with your claim</h2>
+            <h2 class="govuk-heading-s">What happens next</h2>
+            <p class="govuk-body">The Rural Payments Agency (RPA) may contact you to request the review summary the vet has given you.</p>
+      
+            <h2 class="govuk-heading-s">Get help with your claim</h2>
             <p class="govuk-body">Contact the RPA if you need to change the details of your claim.</p>
               
-
             <p class="govuk-body">Have your agreement number with you when you call or include it in your email.</p>
             <ul class="govuk-list">
               <li>Email: <a class="govuk-link" href="mailto:ruralpayments@defra.gov.uk">ruralpayments@defra.gov.uk</a></li>
