+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ffc-sfd-messages",
-  "version": "0.1.17",
+  "version": "0.1.18",
   "description": "",
   "homepage": "github.com?owner=defra&repo=ffc-sfd-messages&organization=defra",
   "main": "app/index.js",
